FInished the settings buttons
Now I have the dropdown aswell, at least idk if the slider also works.

+++ VR Architecture Simulation/Assets/Scripts/Options.cs
@@ -44,7 +44,7 @@ public class Options : UIMenu
     {
         gameObject.SetActive(false);
     }
-    public void ChangeData(byte dataIndex)
+    public void ChangeData(int dataIndex)
     {
         if(dataIndex >= 0 && dataIndex < settingsDataHolder.Count)
         {
